性能优化：（1）将自定义的词典加载，放在项目启动时处理 （2）将核心处理单元配置成bean，全局只new一次 （3）日志级别默认放开到info级别
