+++ src/main/java/com/appleyk/repository/QuestionRepository.java
@@ -1,6 +1,6 @@
 package com.appleyk.repository;
 
-import com.appleyk.node.Movie;
+import com.appleyk.model.Movie;
 import org.springframework.data.neo4j.annotation.Query;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 import org.springframework.data.repository.query.Param;
