+++ src/main/resources/application.properties
@@ -1,21 +1,20 @@
 server.port=8080
-server.servlet.session.timeout=10
 server.tomcat.uri-encoding=utf8
 
-#在application.properties文件中引入日志配置文件
+#\u5728application.properties\u6587\u4EF6\u4E2D\u5F15\u5165\u65E5\u5FD7\u914D\u7F6E\u6587\u4EF6
 #=====================================  log  =============================
 logging.config=classpath:logback-boot.xml
 
-#Neo4j配置
+#Neo4j\u914D\u7F6E
 spring.data.neo4j.username=neo4j
 spring.data.neo4j.password=n123
-#数据库uri地址 
+#\u6570\u636E\u5E93uri\u5730\u5740 
 spring.data.neo4j.uri=bolt://localhost:7687
 
-#HanLP分词字典及自定义问题模板根目录
+#HanLP\u5206\u8BCD\u5B57\u5178\u53CA\u81EA\u5B9A\u4E49\u95EE\u9898\u6A21\u677F\u6839\u76EE\u5F55
 rootDirPath=D:/HanLP/data
 
-#HanLP 用户自定义扩展词库，不建议使用HanLP自定义词典追加的模式，建议自行加载
+#HanLP \u7528\u6237\u81EA\u5B9A\u4E49\u6269\u5C55\u8BCD\u5E93\uFF0C\u4E0D\u5EFA\u8BAE\u4F7F\u7528HanLP\u81EA\u5B9A\u4E49\u8BCD\u5178\u8FFD\u52A0\u7684\u6A21\u5F0F\uFF0C\u5EFA\u8BAE\u81EA\u884C\u52A0\u8F7D
 HanLP.CustomDictionary.path.movieDict=${rootDirPath}/dictionary/custom/movieDict.txt
 HanLP.CustomDictionary.path.genreDict=${rootDirPath}/dictionary/custom/genreDict.txt
 HanLP.CustomDictionary.path.scoreDict=${rootDirPath}/dictionary/custom/scoreDict.txt
