+++ src/main/resources/logback-boot.xml
@@ -40,13 +40,13 @@
     </appender>    
     
     <!-- 控制台输出日志级别 -->
-    <root  level="error">
+    <root  level="info">
           <appender-ref ref="STDOUT" />
      </root> 
     <!-- 指定项目中某个包，当有日志操作行为时的日志记录级别 -->
     <!-- com.appley为根包，也就是只要是发生在这个根包下面的所有日志操作行为的权限都是DEBUG -->
     <!-- 级别依次为【从高到低】：FATAL > ERROR > WARN > INFO > DEBUG > TRACE  --> 
-     <logger name="com.appleyk" level="error">    
+     <logger name="com.appleyk" level="info">
         <appender-ref ref="syslog" />    
     </logger>
