+++ src/test/java/HanLPTest.java
@@ -11,7 +11,7 @@ import org.apache.spark.mllib.linalg.Vectors;
 import org.apache.spark.mllib.regression.LabeledPoint;
 import org.junit.Test;
 
-import com.appleyk.process.ModelProcess;
+import com.appleyk.core.ModelProcess;
 import com.hankcs.hanlp.HanLP;
 import com.hankcs.hanlp.dictionary.CustomDictionary;
 import com.hankcs.hanlp.seg.Segment;
